refactor: signin form and endpoint
